--- android/app/src/main/java/com/example/clarity/PracticeActivity.kt
@@ -1,4 +1,0 @@
-package com.example.clarity
-
-class PracticeActivity {
-}
