--- android/app/src/main/java/com/example/clarity/TestActivity.kt
@@ -1,4 +1,0 @@
-package com.example.clarity
-
-class TestActivity {
-}
