+++ android/app/src/main/java/com/example/clarity/setspage/Card.kt
@@ -1,0 +1,5 @@
+package com.example.clarity.setspage
+
+data class Card(
+    val phrase: String
+)
