+++ android/app/src/main/java/com/example/clarity/setspage/Set.kt
@@ -1,0 +1,8 @@
+package com.example.clarity.setspage
+
+data class Set(
+    val title: String,
+    val cardCount: Int,
+    val cards: MutableList<Card>,
+    val progress: Float,
+)
