+++ android/app/src/main/java/com/example/clarity/setspage/SetsController.kt
@@ -1,0 +1,6 @@
+package com.example.clarity.setspage
+
+class SetsController {
+    
+
+}
