+++ android/app/src/main/java/com/example/clarity/setspage/SetsFragment.kt
@@ -1,10 +1,11 @@
-package com.example.clarity
+package com.example.clarity.setspage
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.clarity.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
