+++ android/app/src/main/res/layout/fragment_sets.xml
@@ -5,12 +5,7 @@
     android:layout_height="match_parent"
     tools:context=".SetsFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Sets Page"
-        android:textSize="26dp"
-        android:layout_gravity="center"/>
-
+    <androidx.recyclerview.widget.RecyclerView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 </FrameLayout>
